- fixed OCL operation and derived attributes, enabled operation reflection
Use a fully qualified class name, when referring to a class from the rtsc
package. Otherwise acceleo throw an exception when evaluation the derived
attribute/operation. This is probably (just a _guess_) due to the fact
that the psm metamodel also registers a "realtimestatechart" package.
Thus, the ocl evaluator cannot disambiguate the relative
"realtimestatechart::<class name>" references.