[Allocation] added EDF and RTA scheduling analyses
Note: the RTA cannot be integrated into the OCLContext.ocl yet due to
beo#477283. The EDF analysis could be merged into the OCLContext.ocl,
though (the question is, if we should modularize OCLContext.ocl a bit
(it's already quite bloated...)).

Both analyses are implemented as described in
"Modeling and efficient solving of extra-functional properties for
adaptation in networked embedded real-time systems adaptation in
networked embedded real-time systems" (Zeller, Prehofer)