[Modelica] fixed rtsc2graphviz transformation
A mapping that declares more than one result returns a tuple and only
for this tuple a trace "record" is established. It seems that prior qvto
version(s) established trace records for the individual results as well
(because this transformation worked some time ago;) ).
(From a quick glance at the 1.2 spec, the current behavior seems to be
reasonable.)