[Allocation] introduced MarkupHoverFragmentFixup to fix the jenkins build
Background:
the MarkupHoverFragment binds the ITextHover to
org.eclipse.ocl.xtext.markup.ui.hover.MarkupCompositeHover, which is only
available in the ocl git repo (since this is just usability, it isn't
worth to backport it...). For now, we just stick to the default xtext binding
(DefaultCompositeHover).

This is just temporary... (at least it's supposed to be temporary;) ).