[Codegen_C] Ignore target folders. Seperate Arduino stuff from rest.
Codegen_C and Codegen_C_Arduino need to be built seperately, because one needs 
a jar of the other. The idea is to set up two jobs that both work in the same folder 
and set one as a repository for the other. 

Maven-generated target folders are ignored.

Source-feature generation: Plugins are added, but it does not work yet. We could not 
figure out why yet.