Implements createDatabase and add main for testing database
Database works so far, storing and retrieving works. I removed passing
the database pointer to all the functions, since that would mean the noSQL Stub
would need to know about unqlite. Added one database pointer to the entire 
class. I think we need to think about how we manage transactions. 

Delete function was not working, because it could not delete an order if there
was not production station for it. Fixed it.

Changed command line output to german. Removed comments that claimed to 
do stuff they didn't-