[ASL] Major wizard code refactoring
The AllocationComputationWizard and EvaluationContextWizard wizards do
not depend on the export wizard anymore. However, the wizard pages in
the *context.muml plugin still inherit from the
AbstractFujabaExportSourcePage, so that muml users benefit from the
"nice/good looking" wizard pages.

The AllocationComputationStrategyConfigurationPage is replaced by the
less "powerful" AllocationComputationStrategyConfigurationWizardPage,
which is sufficient for our needs.